test: variable renaming and move test into other test class
Remove the AdHocSubProcessJobWorkerTest class as it had
only one test
