--- zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/adhocsubprocess/AdHocSubProcessJobWorkerTest.java
@@ -1,106 +1,0 @@
-package io.camunda.zeebe.engine.processing.adhocsubprocess;
-
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import static org.assertj.core.api.Assertions.tuple;
-
-import io.camunda.zeebe.engine.util.EngineRule;
-import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
-import io.camunda.zeebe.model.bpmn.builder.AdHocSubProcessBuilder;
-import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeAdHocImplementationType;
-import io.camunda.zeebe.protocol.impl.record.value.job.JobResult;
-import io.camunda.zeebe.protocol.record.Record;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import io.camunda.zeebe.protocol.record.value.VariableRecordValue;
-import io.camunda.zeebe.test.util.record.RecordingExporter;
-import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
-import java.util.Map;
-import java.util.function.Consumer;
-import org.assertj.core.api.Assertions;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-
-public class AdHocSubProcessJobWorkerTest {
-
-  @ClassRule public static final EngineRule ENGINE = EngineRule.singlePartition();
-
-  private static final String PROCESS_ID = "process";
-  private static final String AD_HOC_SUB_PROCESS_ELEMENT_ID = "ad-hoc";
-  private static final String JOB_TYPE = "routing-agent";
-
-  @Rule public final RecordingExporterTestWatcher watcher = new RecordingExporterTestWatcher();
-
-  private BpmnModelInstance process(final Consumer<AdHocSubProcessBuilder> modifier) {
-    return Bpmn.createExecutableProcess(PROCESS_ID)
-        .startEvent()
-        .adHocSubProcess(
-            AD_HOC_SUB_PROCESS_ELEMENT_ID,
-            adHocSubProcess -> {
-              adHocSubProcess
-                  .zeebeImplementation(ZeebeAdHocImplementationType.JOB_WORKER)
-                  .zeebeJobType(JOB_TYPE);
-              modifier.accept(adHocSubProcess);
-            })
-        .endEvent()
-        .done();
-  }
-
-  @Test
-  public void shouldSetVariableOnCompletingAdHocSubProcess() {
-    // given
-    final BpmnModelInstance process =
-        process(
-            adHocSubProcess -> {
-              adHocSubProcess.zeebeInputExpression("1", "b");
-
-              adHocSubProcess.task("A");
-              adHocSubProcess.task("B");
-              adHocSubProcess.task("C");
-            });
-
-    ENGINE.deployment().withXmlResource(process).deploy();
-
-    final long processInstanceKey =
-        ENGINE.processInstance().ofBpmnProcessId(PROCESS_ID).withVariable("a", 1).create();
-
-    // when
-    final JobResult adHocSubProcess = new JobResult();
-    adHocSubProcess.setCompletionConditionFulfilled(true);
-
-    ENGINE
-        .job()
-        .ofInstance(processInstanceKey)
-        .withType(JOB_TYPE)
-        .withVariables(Map.of("a", 2, "b", 2, "c", 2))
-        .withResult(adHocSubProcess)
-        .complete();
-
-    // then
-    final long adHocSubProcessInstanceKey =
-        RecordingExporter.processInstanceRecords()
-            .withProcessInstanceKey(processInstanceKey)
-            .withElementType(BpmnElementType.AD_HOC_SUB_PROCESS)
-            .getFirst()
-            .getKey();
-
-    Assertions.assertThat(
-            RecordingExporter.variableRecords().withProcessInstanceKey(processInstanceKey).limit(6))
-        .extracting(Record::getValue)
-        .extracting(
-            VariableRecordValue::getName,
-            VariableRecordValue::getValue,
-            VariableRecordValue::getScopeKey)
-        .contains(
-            tuple("a", "2", processInstanceKey),
-            tuple("b", "2", adHocSubProcessInstanceKey),
-            tuple("c", "2", processInstanceKey));
-  }
-}
